- Add getter to CharacterController2D to retrieve "grounded" status of player - Add ClampToScreen utility to allow for keeping a game object inside camera bounds - Modify PlayerMovementController to move player along with scrolling ground - Refactor ground recycle code out of MainController into its own component script, GroundRecycleManager - Add controller script for managing ground movement
