+++ Assets/Scripts/Character/ClampToScreen.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e6aa6f342d9f6a449a43ff3f1cbadd80
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
