+++ Assets/Scripts/Character/PlayerMovementController.cs
@@ -1,5 +1,3 @@
-using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
 
 [RequireComponent(typeof(CharacterController2D))]
@@ -8,14 +6,12 @@ public class PlayerMovementController : MonoBehaviour
 	public float Speed = 50f;
 
 	private CharacterController2D controller;
-	private Rigidbody2D rigidBody;
 
 	#region MonoBehaviour methods
 
 	void Start()
 	{
 		controller = GetComponent<CharacterController2D>();
-		rigidBody = GetComponent<Rigidbody2D>();
 	} //end method Start
 
 	private void Update()
@@ -27,8 +23,8 @@ public class PlayerMovementController : MonoBehaviour
 			controller.Move(Speed * Time.fixedDeltaTime);
 		else if (Input.GetKey(KeyCode.A))
 			controller.Move(-Speed * Time.fixedDeltaTime);
-		else
-			controller.Move(0);
+		else if (controller.Grounded)
+			controller.Move(-28 * Time.fixedDeltaTime, false);
 	} //end method Update
 
 	#endregion
