+++ Assets/Scripts/Scenery.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 00dded7150af6514fa97c823bb9e9e10
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
